code refactoring for effective inlining of scap_next function. Required to split scap-int in 2 separate files and to set m_file in scap_t struct as void*
