+++ userspace/libscap/scap.c
@@ -340,7 +340,7 @@ scap_t* scap_open_offline_int(const char* fname,
 	//
 	// Open the file
 	//
-	handle->m_file = gzopen(fname, "rb");
+	handle->m_file = VD(gzopen(fname, "rb"));
 	if(handle->m_file == NULL)
 	{
 		snprintf(error, SCAP_LASTERR_SIZE, "can't open file %s", fname);
@@ -351,7 +351,7 @@ scap_t* scap_open_offline_int(const char* fname,
 	//
 	// Validate the file and load the non-event blocks
 	//
-	if(scap_read_init(handle, handle->m_file) != SCAP_SUCCESS)
+	if(scap_read_init(handle, GZ(handle->m_file)) != SCAP_SUCCESS)
 	{
 		snprintf(error, SCAP_LASTERR_SIZE, "%s", scap_getlasterr(handle));
 		scap_close(handle);
@@ -410,7 +410,7 @@ void scap_close(scap_t* handle)
 {
 	if(handle->m_file)
 	{
-		gzclose(handle->m_file);
+		gzclose(GZ(handle->m_file));
 	}
 	else
 	{
@@ -503,13 +503,6 @@ uint32_t scap_get_ndevs(scap_t* handle)
 {
 	return handle->m_ndevs;
 }
-
-#ifndef HAS_EARLY_FILTERING
-int32_t scap_next(scap_t* handle, OUT scap_evt** pevent, OUT uint16_t* pcpuid)
-{
-	return scap_next_inline(handle, pevent, pcpuid);
-}
-#endif
 
 //
 // Return the process list for the given handle
