+++ userspace/libscap/scap_next.h
@@ -7,7 +7,7 @@ extern "C" {
 #include "settings.h"
 #include <unistd.h>
 #include "scap.h"
-#include "scap-int.h"
+#include "scap-int_live.h"
 
 #include "../../driver/ppm_ringbuffer.h"
 
@@ -253,7 +253,21 @@ static int32_t scap_next_live(scap_t* handle, OUT scap_evt** pevent, OUT uint16_
 #endif
 }
 
-static __always_inline int32_t scap_next_inline(scap_t* handle, OUT scap_evt** pevent, OUT uint16_t* pcpuid)
+
+/*!
+  \brief Get the next event from the from the given capture instance
+
+  \param handle Handle to the capture instance.
+  \param pevent User-provided event pointer that will be initialized with address of the event.
+  \param pcpuid User-provided event pointer that will be initialized with the ID if the CPU
+    where the event was captured.
+			
+  \return SCAP_SUCCESS if the call is succesful and pevent and pcpuid contain valid data.
+   SCAP_TIMEOUT in case the read timeout expired and no event is available.
+   SCAP_EOF when the end of an offline capture is reached.
+   On Failure, SCAP_FAILURE is returned and scap_getlasterr() can be used to obtain the cause of the error. 
+*/
+static __always_inline int32_t scap_next(scap_t* handle, OUT scap_evt** pevent, OUT uint16_t* pcpuid)
 {
 	int32_t res;
